feat パッケージ中身を個別ページに移動
