+++ backend/app/Http/Controllers/statistics/NamedEntityStatisticsController.php
@@ -5,10 +5,10 @@ namespace App\Http\Controllers\statistics;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\CustomNER;
-use App\Models\NERLabel;
 use App\Models\PackageNER;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Carbon;
+use Illuminate\Routing\Redirector;
+use Illuminate\Http\RedirectResponse;
 
 class NamedEntityStatisticsController extends Controller
 {
@@ -75,7 +75,7 @@ class NamedEntityStatisticsController extends Controller
     // パッケージNERのCRUD周り
 
 
-    public function packagesCreate(Request $request)
+    public function packagesCreate(Request $request): Redirector|RedirectResponse
     {
 
         // バリデーション
@@ -83,15 +83,16 @@ class NamedEntityStatisticsController extends Controller
 
         //中身作成
         $form = [
-            "package_id" => $request->package_id,
+            "package_id" => $request->packageId,
             "label_id" => $request->label_id,
             "name" => $request->name,
         ];
 
         PackageNER::create($form);
-        return redirect('administrator/package#nerTable');
+        return redirect('administrator/package/' . $request->packageId);
     }
-    public function packagesUpdate(Request $request)
+
+    public function packagesUpdate(Request $request): Redirector|RedirectResponse
     {
         // 日付のバリデーション→既に存在する日付ならエラー返す
         // バリデーション
@@ -103,11 +104,12 @@ class NamedEntityStatisticsController extends Controller
         ];
 
         PackageNER::where('id', $request->PackageNER_id)->update($updateContent);
-        return redirect('administrator/package#nerTable');
+        return redirect('administrator/package/' . $request->packageId);
     }
-    public function packagesDelete(Request $request)
+
+    public function packagesDelete(Request $request): Redirector|RedirectResponse
     {
         PackageNER::where('id', $request->PackageNER_id)->delete();
-        return redirect('administrator/package#nerTable');
+        return redirect('administrator/package/' . $request->packageId);
     }
 }
